+++ public/assets/scripts/choices.search-prefix.js
@@ -986,6 +986,7 @@
         prependValue: null,
         appendValue: null,
         renderSelectedChoices: 'auto',
+        searchRenderSelectedChoices: 'auto',
         loadingText: 'Loading...',
         noResultsText: 'No results found',
         noChoicesText: 'No choices to choose from',
@@ -1757,6 +1758,9 @@
             if (typeof config.renderSelectedChoices !== 'boolean') {
                 config.renderSelectedChoices = config.renderSelectedChoices === 'always' || isSelectOne;
             }
+            if (typeof config.searchRenderSelectedChoices !== 'boolean') {
+                config.searchRenderSelectedChoices = config.searchRenderSelectedChoices === 'always' || isSelectOne;
+            }
             if (config.closeDropdownOnSelect === 'auto') {
                 config.closeDropdownOnSelect = isText || isSelectOne || config.singleModeForMultiSelect;
             }
@@ -2406,7 +2410,10 @@
             var fragment = document.createDocumentFragment();
             var renderableChoices = function (choices) {
                 return choices.filter(function (choice) {
-                    return !choice.placeholder && (isSearching ? !!choice.rank : config.renderSelectedChoices || !choice.selected);
+                    return !choice.placeholder &&
+                        (isSearching
+                            ? (config.searchRenderSelectedChoices || !choice.selected) && !!choice.rank
+                            : config.renderSelectedChoices || !choice.selected);
                 });
             };
             var showLabel = config.appendGroupInSearch && isSearching;
@@ -2614,7 +2621,7 @@
             if (!items.length || !this.config.removeItems || !this.config.removeItemButton) {
                 return;
             }
-            var id = element && parseDataSetId(element.parentElement);
+            var id = element && parseDataSetId(element.closest('[data-id]'));
             var itemToRemove = id && items.find(function (item) { return item.id === id; });
             if (!itemToRemove) {
                 return;
@@ -3194,7 +3201,7 @@
          */
         Choices.prototype._onMouseDown = function (event) {
             var target = event.target;
-            if (!(target instanceof HTMLElement)) {
+            if (!(target instanceof Element)) {
                 return;
             }
             // If we have our mouse down on the scrollbar and are on IE11...